# feat(outputs.sql): Use Prepared statements for write tx