## Summary

The current approach to writing, issues an INSERT query for every metric in the buffer, which while working, leaves room for improvement.

This commit modifies the Write behavior as follows:

- Sort the metrics fields/tags (to ensure greater cache-hit ratio)
- Use Prepared statements to write multiple metrics with a single query

The results depends on the metrics buffer, the metric cardinality, the output server configuration, etc.

That being said, in micro-benchmarks and production deployment, we've seen 10-100x improvement in write performance.

### Benchmark:

Batches of 1000, 5000, 10000, 50000 and 100000 metrics.

```
│                                        │     Write    │              Prepared               │
                                         │    sec/op    │   sec/op     vs base                │
SqliteThroughput/metrics_insert_1000-4     93.199m ± 0%   7.722m ± 2%  -91.71% (p=0.000 n=10)
SqliteThroughput/metrics_insert_5000-4     466.31m ± 0%   37.40m ± 1%  -91.98% (p=0.000 n=10)
SqliteThroughput/metrics_insert_10000-4    931.11m ± 0%   74.89m ± 1%  -91.96% (p=0.000 n=10)
SqliteThroughput/metrics_insert_50000-4    4663.6m ± 0%   374.4m ± 0%  -91.97% (p=0.000 n=10)
SqliteThroughput/metrics_insert_100000-4   9323.9m ± 0%   749.2m ± 1%  -91.96% (p=0.000 n=10)
geomean                                      1.120        90.49m       -91.92%
```

Note:

There are many flaws with this benchmark: SQLite is not representative of every driver, system-specific conditions, metrics had very low cardinality (fewer INSERT queries).

But while a bad benchmark, it's not completely useless and it shows the stark difference in extreme cases.

## Checklist

- [X] No AI generated code was used in this PR