@hautecodure 2025-04-14
Thanks for taking the time to review the PR.

> Could you please add an option to enable prepared statements (defaulting to using non-prepared statements) and try to minimize the code restructuring!

Do you find it acceptable to bring the previous version as another function, something like:

```go
//...
func (p *SQL) sendSingleTx(sql string) err {

}
//...
		// sorting improves the Prepared statement throughput, when
		// fields/tags arrive in different order
		sortedColumns, sortedParams := p.sortFields(columns, values)

		sql := p.generateInsert(tablename, sortedColumns)

		if p.UseBatchedTx {
                  batchedQueries[sql] = append(batchedQueries[sql], sortedParams)
		} else {
                  if err := sendSingleTx(sql); err != nil {
                    return err
                  }
		}
	}

        if p.UseBatchedTx {
          for sql, sqlParams := range batchedQueries {
            if err := p.sendBatchTx(sql, sqlParams); err != nil {
              return fmt.Errorf("failed to send a batched tx: %w", err)
            }
          }
        }

	return nil
}
```

While the default seem "hidden" (behind the else condition), i think it's more readable as both if conditions (if p.UseBatchedTx) are the same.

I think you mentioned it's easier to debug, having the code in one place, but i think it would unreadable if i start cascading if/else around the existing version