@srebhan 2025-04-17
@hautecodure yeah, move the current code into an e.g. `sendUnprepared` function  and move the new code to an `sendPrepared` function. Make sure that the _old_ code-path is the same as before.

Another idea. Why not push your approach further? Add a config option
```toml
  ## Send metrics with the same columns and the same table as batches using prepared statements
  # batch_send = false
``` 

Keep the old code sending the data directly (including the clickhouse exception) if `batch_send` is `false` (default). In case the user wants batch-sends, create a plugin-level cache for caching the prepared SQL statements like `map[uint64]*sql.Stmt`. The key should be the hash and the parameters the columns. Like returned by

```go
func processMetric(m telegraf.metric) (uint64,  []interface{}) {
	// Prepare the hash
	h := fnv.New64a()
	h.Write([]byte(m.MetricName()))
	h.Write([]byte("\n"))

	// Prepare the parameters
	params := make([]interface{}, 0, 1+len(tags)+len(fields))
	params = append(params, m.Time())
	
	// Tags are already sorted but fields need to be sorted to end-up in the right spot
	fields := make([]*telegraf.Field, len(m.FieldList()))
	copy(fields, m.FieldList())
	slices.SortFunc(fields, func(a, b telegraf.Field) int { return strings.Compare(a.Key, b.Key) }

	// Add the data to both the parameters and hash
	for _, tag := range m.TagList() {
		h.Write([]byte(tag.Key+"\n"))
		params = append(params, tag.Value)
	}
	for _, field := range fields {
		h.Write([]byte(field.Key+"\n"))
		params = append(params, field.Value)
	}

	return h.Sum64(), params
```