@hautecodure 2025-04-17
Thanks for the input. Just want to clarify:

Do you mean caching the prepared statement in the p struct?

If so, i think it will be troublesome to keep the code simple as the documentation [urges](https://go.dev/doc/database/prepared-statements) to Close() a Statement when not in use, which would introduce state to an otherwise stateless operation - writing.

We could cache the output from generateInsert (instead of gosql.Stmt), but when introducing a hash, I'm not sure it will be a net-benefit, performance-wise. Though i have not run any benchmarks to confirm this.