@srebhan 2025-04-23
Oh didn't know a statement uses resources so yeah you are right, we should not cache the prepared statement(s) across write calls. Anyway, I'm pretty sure computing the hash is faster than recreating the statement so I would appreciate if you could benchmark this.