@hautecodure 2025-06-13
Hi, sorry for responding so late, there was a priority shift the last few months, so we couldn't dedicate as much time to up-streaming this PR.

I ran a benchmark with the new processing logic (as suggested) and it is indeed a bit (~2%) faster:

```
│                                        │     Write    │              Prepared               │
                                         │    sec/op    │   sec/op     vs base                │
SqliteThroughput/metrics_insert_1000-12    96.629m ± 1%   6.301m ± 11%  -93.48% (p=0.000 n=10)
SqliteThroughput/metrics_insert_5000-12    479.83m ± 1%   29.68m ±  3%  -93.82% (p=0.000 n=10)
SqliteThroughput/metrics_insert_5000-12    961.55m ± 1%   58.56m ±  1%  -93.91% (p=0.000 n=10)
SqliteThroughput/metrics_insert_10000-12   4831.4m ± 0%   294.8m ±  2%  -93.90% (p=0.000 n=10)
SqliteThroughput/metrics_insert_50000-12   9672.9m ± 0%   586.4m ±  1%  -93.94% (p=0.000 n=10)
geomean                                     1.158        71.69m        -93.81%
```

I pushed a new version of this PR, with a couple of improvements:

- Hashing (as suggested) insert queries.

In order to avoid duplicating code, i think it's a good idea to use the processMetric function for the existing users as well (i.e. those not using Prepared statements).

Re-ordering (sorting) the query columns will have no impact on existing users and it saves us having to write / maintain two Write paths.

- Use XXHash to hash the keys

While fnv is built-in and present in the stdlib, xxhash is pulled indirectly (so it's not a new dependency) and it supposedly offers better collision resistance and faster performance across multiple platforms.

However, I'm not a cryptographer or an expert in the field of one-time hashing, so I'm not able to provide an in-depth comparison at this time.

Documentation and test are missing (as of now) as i would like to get some feedback on this approach.